+++ Ordenacao/countingSort.cpp
@@ -16,7 +16,7 @@ void CountingSort::ordena(int * vet, int tam)
 {
     this->start_time = chrono::system_clock::now();
 
-    register int instrucoes = 0;
+    register unsigned long long int instrucoes = 0;
 
     instrucoes += 2;
     this->vetAXU = new int[tam + 1];
